+++ src/DotRecast.Detour.Extras/Jumplink/PolyQueryInvoker.cs
@@ -4,16 +4,16 @@ namespace DotRecast.Detour.Extras.Jumplink
 {
     public class PolyQueryInvoker : IDtPolyQuery
     {
-        public readonly Action<DtMeshTile, DtPoly, long> _callback;
+        private readonly Action<DtMeshTile, DtPoly[], long[], int> _callback;
 
-        public PolyQueryInvoker(Action<DtMeshTile, DtPoly, long> callback)
+        public PolyQueryInvoker(Action<DtMeshTile, DtPoly[], long[], int> callback)
         {
             _callback = callback;
         }
 
-        public void Process(DtMeshTile tile, DtPoly poly, long refs)
+        public void Process(DtMeshTile tile, DtPoly[] poly, long[] refs, int count)
         {
-            _callback?.Invoke(tile, poly, refs);
+            _callback?.Invoke(tile, poly, refs, count);
         }
     }
 }